+++ src/main/kotlin/feature/budget/domain/model/Budget.kt
@@ -1,6 +1,7 @@
 package feature.budget.domain.model
 
 import kotlinx.datetime.LocalDate
+import java.math.BigDecimal
 import java.util.*
 
 data class Budget(
@@ -8,7 +9,7 @@ data class Budget(
     val accountIds: List<UUID>,
     val name: String,
     val categoryIds: List<UUID>,
-    val limit: Double,
+    val limit: BigDecimal,
     val isExpense: Boolean,
     val startDate: LocalDate,
     val endDate: LocalDate
